tests/wpt/tests/html/semantics/forms/the-selectmenu-element/selectmenu-value-option.tentative.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://github.com/openui/open-ui/issues/664">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<selectmenu id=selectmenu>
  <option id=optone>innertext one</option>
  <option id=opttwo value=valueattribute>innertext two</option>
</selectmenu>

<script>
test(() => {
  assert_equals(selectmenu.value, 'innertext one',
    'The first option should be selected initially.');
  selectmenu.value = 'valueattribute';
  assert_equals(selectmenu.value, 'valueattribute',
    'Assigning value should look at the options value, not innertext');
}, 'selectmenu.value should reflect option.value');
</script>
